refactor Commands.ts
