+++ src/classes/Commands/export.ts
@@ -1,0 +1,9 @@
+import * as botStatus from './botStatusCommands';
+import * as help from './helpCommands';
+import messageCommand from './messageCommand';
+import * as misc from './miscCommands';
+import * as pricelist from './pricelistManagerCommands';
+import * as review from './reviewCommands';
+import * as utils from './utils';
+
+export { botStatus, help, messageCommand, misc, pricelist, review, utils };
